api/requirements.txt
# ============================================================
# API Server Dependencies
# ============================================================

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# GraphQL
strawberry-graphql[fastapi]>=0.215.0

# Database
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
redis>=5.0.0

# Kafka
aiokafka>=0.9.0
kafka-python>=2.0.2

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# WebSocket
websockets>=12.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Monitoring
prometheus-client>=0.19.0

# CORS
starlette>=0.27.0
